+++ ManuscriptPipeline/Models/BibliographyFormat.vb
@@ -1,0 +1,6 @@
+Namespace Models
+    Public Enum BibliographyFormat
+        BibTeX
+        Ris
+    End Enum
+End Namespace