+++ ManuscriptPipeline/Models/BibliographyImportOptions.vb
@@ -1,0 +1,7 @@
+Namespace Models
+
+    Public Class BibliographyImportOptions
+        Public Property ImportPublishedRecordsAsPublished As Boolean = True
+    End Class
+
+End Namespace