+++ ManuscriptPipeline/Models/BibliographyImportResult.vb
@@ -1,0 +1,10 @@
+Namespace Models
+
+    Public Class BibliographyImportResult
+        Public Property ImportedCount As Integer
+        Public Property DuplicateCount As Integer
+        Public Property AuthorsCreated As Integer
+        Public Property WarningCount As Integer
+    End Class
+
+End Namespace